separate logic for websocket managing to another file
