+++ kono-judge-status/src/components/websocket/WebSocketController.js
@@ -1,6 +1,18 @@
-import React, { useState } from 'react';
+import React, { useEffect } from 'react';
+import { useWebSocket } from 'lib/hooks';
 
 export default () => {
+
+    const [register, unregister] = useWebSocket();
+
+    console.log(register, unregister);
+
+    // useEffect(() => {
+    //     const eventHandler = register();
+    //     return function cleanup() {
+    //         unregister(eventHandler);
+    //     }
+    // }, [register, unregister]);
 
     return (
         <div>
